Move portfolio simulation to portfolio module
